Reorganize parametric_umap
