+++ src/glass_box_umap/parametric_umap/__init__.py
@@ -1,19 +1,9 @@
-from glass_box_umap.parametric_umap.main import (
+from .api import (
     PUMAP,
-    PeriodicCheckpoint,
-    UMAPDataModule,
-    UMAPLightningModule,
     load_pumap,
 )
-from glass_box_umap.parametric_umap.model import ConvEncoder, DefaultDecoder, DefaultEncoder
 
 __all__ = [
     "PUMAP",
     "load_pumap",
-    "UMAPLightningModule",
-    "UMAPDataModule",
-    "PeriodicCheckpoint",
-    "DefaultEncoder",
-    "DefaultDecoder",
-    "ConvEncoder",
 ]
