+++ src/glass_box_umap/parametric_umap/data.py
@@ -1,52 +1,11 @@
 import numpy as np
 import torch
 from numpy.typing import NDArray
-from scipy.sparse import coo_matrix, csr_matrix
+from scipy.sparse import csr_matrix
 from torch import Tensor
 from torch.utils.data import Dataset
 
-GraphElements = tuple[
-    coo_matrix,
-    NDArray[np.floating],
-    NDArray[np.intp],
-    NDArray[np.intp],
-    NDArray[np.floating],
-    int,
-]
-
-
-def get_graph_elements(graph: csr_matrix, n_epochs: int) -> GraphElements:
-    """Extract graph elements from a sparse UMAP graph for edge sampling.
-
-    Converts a sparse graph representation into arrays of edge indices and weights
-    suitable for training a parametric UMAP model.
-
-    Args:
-        graph: Sparse CSR matrix representing the UMAP graph with edge weights.
-        n_epochs: Number of training epochs, used to determine sampling frequency.
-
-    Returns:
-        A tuple containing:
-            - graph: The COO format graph with low-probability edges removed
-            - epochs_per_sample: Number of times each edge should be sampled per epoch
-            - head: Source vertex indices for each edge
-            - tail: Target vertex indices for each edge
-            - weight: Edge weights
-            - n_vertices: Total number of vertices in the graph
-    """
-    graph_coo = graph.tocoo()
-    graph_coo.sum_duplicates()
-    n_vertices = graph_coo.shape[1]
-
-    graph_coo.data[graph_coo.data < (graph_coo.data.max() / float(n_epochs))] = 0.0
-    graph_coo.eliminate_zeros()
-
-    epochs_per_sample = n_epochs * graph_coo.data
-    head = graph_coo.row
-    tail = graph_coo.col
-    weight = graph_coo.data
-
-    return graph_coo, epochs_per_sample, head, tail, weight, n_vertices
+from .graph import get_graph_elements
 
 
 class UMAPDataset(Dataset[tuple[Tensor, Tensor]]):
@@ -64,7 +23,7 @@ class UMAPDataset(Dataset[tuple[Tensor, Tensor]]):
     def __init__(
         self,
         data: NDArray[np.floating],
-        graph: "csr_matrix",
+        graph: csr_matrix,
         n_epochs: int = 200,
     ) -> None:
         _, epochs_per_sample, head, tail, _, _ = get_graph_elements(graph, n_epochs)
@@ -78,35 +37,9 @@ class UMAPDataset(Dataset[tuple[Tensor, Tensor]]):
         self.data = torch.as_tensor(data, dtype=torch.float32)
 
     def __len__(self) -> int:
-        return int(self.data.shape[0])
+        return len(self.data.shape[0])
 
     def __getitem__(self, index: int) -> tuple[Tensor, Tensor]:
         edges_to_exp = self.data[self.edges_to_exp[index]]
         edges_from_exp = self.data[self.edges_from_exp[index]]
         return edges_to_exp, edges_from_exp
-
-
-class MatchDataset(Dataset[tuple[Tensor, Tensor]]):
-    """PyTorch Dataset for matching parametric embeddings to non-parametric ones.
-
-    Used when training a parametric model to reproduce embeddings from a
-    pre-computed non-parametric UMAP.
-
-    Args:
-        data: Input data array of shape (n_samples, ...).
-        embeddings: Pre-computed UMAP embeddings of shape (n_samples, n_components).
-    """
-
-    def __init__(
-        self,
-        data: NDArray[np.floating],
-        embeddings: NDArray[np.floating],
-    ) -> None:
-        self.data = torch.as_tensor(data, dtype=torch.float32)
-        self.embeddings = torch.as_tensor(embeddings, dtype=torch.float32)
-
-    def __len__(self) -> int:
-        return int(self.data.shape[0])
-
-    def __getitem__(self, index: int) -> tuple[Tensor, Tensor]:
-        return self.data[index], self.embeddings[index]
