+++ src/glass_box_umap/parametric_umap/lightning/data.py
@@ -1,0 +1,34 @@
+import pytorch_lightning as pl
+from torch.utils.data import DataLoader
+
+from ..data import UMAPDataset
+
+
+class UMAPDataModule(pl.LightningDataModule):
+    """PyTorch Lightning data module for UMAP training.
+
+    Args:
+        dataset: PyTorch dataset providing training samples.
+        batch_size: Number of samples per batch.
+        num_workers: Number of worker processes for data loading.
+    """
+
+    def __init__(
+        self,
+        dataset: UMAPDataset,
+        batch_size: int,
+        num_workers: int,
+    ) -> None:
+        super().__init__()
+        self.dataset = dataset
+        self.batch_size = batch_size
+        self.num_workers = num_workers
+
+    def train_dataloader(self) -> DataLoader:
+        return DataLoader(
+            dataset=self.dataset,
+            batch_size=self.batch_size,
+            num_workers=self.num_workers,
+            shuffle=True,
+            persistent_workers=self.num_workers > 0,
+        )
